fix gsheet unit tests
