+++ perfmetrics/scripts/testing_on_gke/examples/fio/run_tests.py
@@ -29,8 +29,8 @@ import sys
 
 # local imports from other directories
 sys.path.append(os.path.join(os.path.dirname(__file__), '..', 'utils'))
-from run_tests_common import escape_commas_in_string, parse_args, run_command, add_iam_role_for_buckets
-from utils import UnknownMachineTypeError, resource_limits
+from run_tests_common import escape_commas_in_string, parse_args, add_iam_role_for_buckets
+from utils import UnknownMachineTypeError, resource_limits, run_command
 
 # local imports from same directory
 import fio_workload
