+++ perfmetrics/scripts/testing_on_gke/examples/utils/run_tests_common.py
@@ -18,21 +18,7 @@
 import argparse
 import subprocess
 import sys
-
-
-def run_command(command: str) -> int:
-  """Runs the given string command as a subprocess.
-
-  Returns exit-code which would be non-zero for error.
-  """
-  result = subprocess.run(
-      [word for word in command.split(' ') if (word and not str.isspace(word))],
-      capture_output=True,
-      text=True,
-  )
-  print(result.stdout)
-  print(result.stderr)
-  return result.returncode
+from utils import run_command
 
 
 def escape_commas_in_string(unescapedStr: str) -> str:
